+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+Nothing yet.
+
 ## 1.1.0
 
 ### Added
@@ -10,7 +14,7 @@ This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a C
 
 ### Fixed
 
-- Incorrect value being passed to `BaseValidator::testValue()` in `BaseValidator::isValid()`. 
+- Incorrect value being passed to `BaseValidator::testValue()` in `BaseValidator::isValid()`.
 
 ## 1.0.0