+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a C
 
 ### Fixed
 
-- `Not` returned error messages before first call to `::isValid()`.
+- `Not::getMessages()` returned failure messages before first call to `::isValid()`.
 
 ## 1.1.0