+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a C
 
 - `Not::getMessages()` returned failure messages before first call to `::isValid()`.
 - `Not::getMessages()` returned an indexed array of messages.
+- `Comparison` and `Type` referenced incorrect failure message keys when validating options.
 
 ## 1.1.0